requirements.txt
openai
pytubefix
python-dotenv
streamlit